src/requirements.txt
google-cloud-bigquery-storage
cupy-cuda110
typing_extensions>=3.10.0
git+https://github.com/jqueguiner/DeOldify.git@master
pillow>=7.1.0
transformers>=4.9.2
uvicorn==0.14.0
fastapi==0.65.2
fastapi_utils==0.2.1
icecream==2.1.0
prometheus_fastapi_instrumentator==5.7.1
python_forge==18.6.0
googledrivedownloader==0.4
python-multipart==0.0.5
redisai==1.0.2
face_recognition==1.3.0
pycocotools==2.0.2
submitit==1.3.3
pytesseract==0.3.8
easyocr==1.4.1
https://download.openmmlab.com/mmcv/dist/1.3.5/torch1.7.0/cu110/mmcv_full-latest%2Btorch1.7.0%2Bcu110-cp37-cp37m-manylinux1_x86_64.whl
mmdet==2.11.0
Polygon3==3.0.9.1
pyclipper==1.3.0
mmcv-full>=1.3.4
mmdet>=2.11.0
imgaug==0.4.0
lanms-proper==1.0.1
lmdb==1.2.1
matplotlib
numba>=0.45.1
numpy
Polygon3
pyclipper==1.3.0
rapidfuzz==1.6.2
scikit-image==0.18.3
six==1.15.0
terminaltables
ucr==0.2.16
sentence-transformers==2.0.0
happytransformer==2.3.1
nltk==3.6.5
sentencepiece==0.1.96
pytorch-crf==0.7.2
seqeval==1.2.2
keybert==0.4.0
inflect==5.3.0
einops==0.3.2
nop==0.1.1
cog==0.0.3
aiofiles==0.7.0
LanguageIdentifier==1.1.1
orjson==3.6.4
git+https://github.com/ramsrigouthamg/Questgen.ai
git+https://github.com/boudinfl/pke.git
git+https://github.com/theunifai/unifai-api-utils.git
